+++ FluentFTP/Helpers/Parsers/FtpWindowsParser.cs
@@ -19,7 +19,7 @@ namespace FluentFTP.Helpers.Parsers {
 		/// <param name="record">A line from the listing</param>
 		/// <param name="capabilities">Server capabilities</param>
 		/// <returns>FtpListItem if the item is able to be parsed</returns>
-		public static FtpListItem ParseLegacy(string record, FtpCapability capabilities, FtpClient client) {
+		public static FtpListItem ParseLegacy(string record, List<FtpCapability> capabilities, FtpClient client) {
 			var item = new FtpListItem();
 			var datefmt = new string[] {
 				"MM-dd-yy  hh:mmtt",